# perf(weave): filter op_name pre group by 